Extracting ArticleList into own component
